requirements.txt
streamlit>=1.30.0
requests>=2.31.0
pillow>=10.0.0
numpy>=1.24.0
torch>=2.2.0
torchvision>=0.17.0
scikit-learn>=1.4.0
joblib>=1.3.0
matplotlib>=3.7.0
